Revert "[logging] WIP > Extract the loggers into separate package (!DRAFT!)"
This reverts commit 90c2126c2f660e766720d907f24e96733013ce07.
